+++ src/main/java/org/itsci/assist_decisions/repository/PostRepository.java
@@ -41,7 +41,7 @@ public interface PostRepository extends JpaRepository<Post,String> {
     @Transactional
     @Modifying
     @Query(value = "UPDATE post SET result = :result , avgPoint = postPoint/:amount WHERE postId = :postId", nativeQuery = true)
-    void updateResult(@Param("result") String result ,@Param("postId")String postId,Double amount);
+    void updateResult(@Param("result") String result ,@Param("postId")String postId,int amount);
 
 
     @Query(value = "SELECT p.* FROM post p JOIN choice c ON p.postId = c.postId JOIN vote v ON c.choiceId = v.choiceId JOIN member m ON v.username = m.username WHERE m.username = :username and result != 'r' ORDER BY v.voteDate", nativeQuery = true)
